requirements.txt
pyusb>=1.0.0
pyinstaller>=6.0.0